Improving the directory structure
